tests/test__ebml.py

import io

from tests import TestCase, add
from mutagen._compat import long_
from mutagen._ebml import (Element, UnsignedElement, ASCIIElement, UTF8Element,
                           BinaryElement, MasterElement, EBMLHeader)


class TEBMLElement(TestCase):

    def test_read_id(self):
        one = io.BytesIO(b'\xBF')
        two = io.BytesIO(b'\x42\x85')
        three = io.BytesIO(b'\x3C\xB9\x23')
        four = io.BytesIO(b'\x1A\x45\xDF\xA3')

        self.assertEqual(Element._read_id(one), 0xBF)
        self.assertEqual(Element._read_id(two), 0x4285)
        self.assertEqual(Element._read_id(three), 0x3CB923)
        self.assertEqual(Element._read_id(four), 0x1A45DFA3)

    def test_write_id(self):
        test_values = [
            (0xBF, b'\xBF'),
            (0x4285, b'\x42\x85'),
            (0x3CB923, b'\x3C\xB9\x23'),
            (0x1A45DFA3, b'\x1A\x45\xDF\xA3'),
        ]

        for _in, _out in test_values:
            buf = io.BytesIO()
            Element._write_id(buf, _in)
            self.assertEqual(buf.getvalue(), _out)

    def test_id_roundtrip(self):
        buf = io.BytesIO(b'\x1A\x45\xDF\xA3')
        _id = Element._read_id(buf)

        buf = io.BytesIO()
        Element._write_id(buf, _id)

        self.assertEqual(buf.getvalue(), b'\x1A\x45\xDF\xA3')

    def test_size_read(self):
        test_values = [
            (b'\x81', 0x1),
            (b'\x41\xA3', 0x1A3),
            (b'\x25\xB3\xC2', 0x5B3C2),
            (b'\x16\xAE\x7D\x3F', 0x6AE7D3F),
            (b'\x08\xF4\x5E\xCD\x0A', 0xF45ECD0A),
            (b'\x04\xAB\x84\xCB\x69\xFD', 0xAB84CB69FD),
            (b'\x02\xDF\x38\xBA\xE4\x75\xAD', 0xDF38BAE475AD),
            (b'\x01\xE4\x51\xF3\x3E\xB8\xA5\x6D', 0xE451F33EB8A56D),
        ]

        for _in, _out in test_values:
            buf = io.BytesIO(_in)
            self.assertEqual(Element._read_size(buf), _out)

    def test_size_write(self):
        test_values = [
            (0x81, b'\x81'),
            (0x41A3, b'\x41\xA3'),
            (0x25B3C2, b'\x25\xB3\xC2'),
            (0x16AE7D3F, b'\x16\xAE\x7D\x3F'),
            (0x08F45ECD0A, b'\x08\xF4\x5E\xCD\x0A'),
            (0x04AB84CB69FD, b'\x04\xAB\x84\xCB\x69\xFD'),
            (0x02DF38BAE475AD, b'\x02\xDF\x38\xBA\xE4\x75\xAD'),
            (0x01E451F33EB8A56D, b'\x01\xE4\x51\xF3\x3E\xB8\xA5\x6D')
        ]

        for _in, _out in test_values:
            buf = io.BytesIO()
            Element._write_size(buf, _in)
            self.assertEqual(buf.getvalue(), _out)

    def test_element_size(self):
        original_size = 100
        dataobj = io.BytesIO(b'\x00')
        Element._write_size(dataobj, original_size)

        dataobj.seek(0)
        new_size = Element._read_size(dataobj)

        self.assertEqual(original_size, new_size)
        self.assertEqual(dataobj.getvalue(), b'\xe4')

    def test_unsigned(self):
        ele = UnsignedElement(0x1A45DFA3, 100)
        self.assertEqual(ele.bytes_to_write(), 6)

        dataobj = io.BytesIO(b'\x00' * ele.bytes_to_write())
        ele.write(dataobj)

        dataobj.seek(0)

        _id = Element._read_id(dataobj)
        size = Element._read_size(dataobj)

        ele = UnsignedElement.from_buffer(_id, size, dataobj)

        self.assertEqual(ele, 100)
        self.assertEqual(dataobj.getvalue(), b'\x1a\x45\xdf\xa3\x81\x64')

    def test_ascii(self):
        ele = ASCIIElement(0x1A45DFA3, u'abcd')
        self.assertEqual(ele.bytes_to_write(), 9)

        dataobj = io.BytesIO(b'\x00' * ele.bytes_to_write())
        ele.write(dataobj)

        dataobj.seek(0)

        _id = Element._read_id(dataobj)
        size = Element._read_size(dataobj)

        ele = ASCIIElement.from_buffer(_id, size, dataobj)

        self.assertEqual(ele, u'abcd')
        self.assertEqual(dataobj.getvalue(), b'\x1a\x45\xdf\xa3\x84abcd')

    def test_utf8(self):
        ele = UTF8Element(0x1A45DFA3, u'Test\u2019s')

        self.assertEqual(ele.bytes_to_write(), 13)

        dataobj = io.BytesIO(b'\x00' * ele.bytes_to_write())
        ele.write(dataobj)

        dataobj.seek(0)

        _id = Element._read_id(dataobj)
        size = Element._read_size(dataobj)

        ele = UTF8Element.from_buffer(_id, size, dataobj)

        self.assertEqual(ele, u'Test\u2019s')
        self.assertEqual(dataobj.getvalue(),
                         b'\x1a\x45\xdf\xa3\x88Test\xe2\x80\x99s')

    def test_binary(self):
        ele = BinaryElement(0x1A45DFA3, b'\x00\x11\x22\x33')

        self.assertEqual(ele.bytes_to_write(), 9)

        dataobj = io.BytesIO(b'\x00' * ele.bytes_to_write())
        ele.write(dataobj)

        dataobj.seek(0)

        _id = Element._read_id(dataobj)
        size = Element._read_size(dataobj)

        ele = BinaryElement.from_buffer(_id, size, dataobj)

        self.assertEqual(ele, b'\x00\x11\x22\x33')
        self.assertEqual(dataobj.getvalue(),
                         b'\x1a\x45\xdf\xa3\x84\x00\x11\x22\x33')

    def test_master(self):
        ele = MasterElement(0x1A45DFA3)
        ele._children = [
            BinaryElement(0x1A45DFA3, b'\x00\x11\x22\x33'),
            UTF8Element(0x1A45DFA3, u'Test\u2019s'),
            ASCIIElement(0x1A45DFA3, u'abcd'),
            UnsignedElement(0x1A45DFA3, 100),
        ]
        # Adding elements like this isn't actually supported, so set bytes_read
        # to compensate
        ele.bytes_read = 5

        self.assertEqual(ele.data_size, 37)
        self.assertEqual(ele.bytes_to_write(), 42)

        dataobj = io.BytesIO(b'\x00' * ele.bytes_to_write())
        ele.write(dataobj)

        dataobj.seek(0)

        self.assertEqual(
            dataobj.getvalue(),
            b'\x1a\x45\xdf\xa3\xa5\x1a\x45\xdf\xa3\x84\x00\x11\x22\x33\x1a\x45'
            b'\xdf\xa3\x88Test\xe2\x80\x99s\x1a\x45\xdf\xa3\x84abcd\x1a\x45'
            b'\xdf\xa3\x81\x64'
        )

    def test_ebml_header(self):
        data = (b'\x1A\x45\xDF\xA3\x01\x00\x00\x00\x00\x00\x00\x23\x42\x86\x81'
                b'\x01\x42\xF7\x81\x01\x42\xF2\x81\x04\x42\xF3\x81\x08\x42\x82'
                b'\x88\x6D\x61\x74\x72\x6F\x73\x6B\x61\x42\x87\x81\x02\x42\x85'
                b'\x81\x02')
        dataobj = io.BytesIO(data)

        _id = Element._read_id(dataobj)
        size = Element._read_size(dataobj)
        ele = EBMLHeader.from_buffer(_id, size, dataobj)

        self.assertEquals(ele.data_size, 35)
        # This currently fails - need to work out how to deal with files using
        # more than minimum space
        # self.assertEquals(ele.bytes_read, 35)
        # self.assertEquals(ele._get_byte_difference(), 0)
        self.assertEquals(ele.bytes_to_write(), 40)

        dataobj = io.BytesIO(b'\x00' * ele.bytes_to_write())
        ele.write(dataobj)

        #self.assertEquals(dataobj.getvalue(), data)

    def test_document(self):
        pass


add(TEBMLElement)
